feat(timer): add execution time measurement for event retrieval (master)
This commit introduces a decorator to measure and log the execution
time of the `get_events` function.

By wrapping the function with the `timer` decorator, developers
can now easily track the performance of the event retrieval process.
This enhancement will aid in identifying potential bottlenecks
and optimizing the code in the future.

Signed-off-by: furterg <gregory.furter@gmail.com>