fix: correct calculation of duration days in 'Durée_first_char' column.
This change modifies the way the number of days is extracted from the
'Durée' column. The previous implementation subtracted one from the
extracted value, which was incorrect.

Now, the code accurately reflects the intended duration by directly
using the extracted integer value. This ensures that the subsequent
calculation of the 'End Date' column is based on the correct number
of days.

This fix improves the accuracy of date calculations in the data
processing workflow. (master)

Signed-off-by: furterg <gregory.furter@gmail.com>