fix: correct calculation of duration days in DataFrame (master)
This commit fixes an issue in the calculation of the number of days
extracted from the 'Durée' column in the DataFrame.

The previous implementation incorrectly subtracted 1 from the
extracted integer during the conversion, which could lead to
inaccurate results.

The updated code ensures that the number of days is accurately
calculated, thereby improving the reliability of the 'End Date'
column generation.

Signed-off-by: furterg <gregory.furter@gmail.com>